fix condition

+++ models/engine/db_storage.py
@@ -25,8 +25,8 @@ class DBStorage:
             getenv("HBNB_MYSQL_HOST"),
             getenv("HBNB_MYSQL_DB")), pool_pre_ping=True)
 
-        # if getenv("HBNB_ENV") == "test":
-        Base.metadata.drop_all(self.__engine)
+        if getenv("HBNB_ENV") == "test":
+            Base.metadata.drop_all(self.__engine)
 
     def all(self, cls=None):
         """Select data from table"""
